requirements.txt
openai
python-dotenv
pypdf
python-docx
numpy